+++ modules/TFTP.py
@@ -1,5 +1,5 @@
 import argparse
-import os
+import sys
 import subprocess
 
 import tools
@@ -15,19 +15,23 @@ class TFTP(BaseModule):
         self.rootdir = None
         self.pidfile = None
         self.process = None
+        self.enabled = True
 
         # This module requires tftpd
         self.binary = tools.locate('in.tftpd')
         if self.binary is None:
-            raise FileNotFoundError('The TFTP module requires tftpd to be installed and on $PATH.')
+            print('The TFTP module requires tftpd to be installed and on $PATH.', file=sys.stderr)
+            self.enabled = False
 
         try:
             self.version = subprocess.check_output(['tftpd', '--version'],
                                                    stderr=subprocess.STDOUT).decode().strip()
             if not self.version:
-                raise RuntimeError('The TFTP module could not detect tftpd version.')
+                print('The TFTP module could not detect tftpd version.', file=sys.stderr)
+                self.enabled = False
         except:
-            raise RuntimeError('The TFTP module could not run tftpd.')
+            print('The TFTP module could not run tftpd.', file=sys.stderr)
+            self.enabled = False
 
         self.stdout = tools.ThreadOutput('TFTP')
 
@@ -39,10 +43,17 @@ class TFTP(BaseModule):
                             help='Set path for tftpd root directory, default is "/srv/tftp/"')
 
     def configure(self, args):
+        if not self.enabled and args.pxe:
+            print('TFTP module requested for network booting, but it will not run!', file=sys.stderr)
+
         self.rootdir = args.tftp_rootdir
         self.pidfile = args.tftp_pidfile
+        self.enabled_user = self.enabled and args.pxe
 
     def start(self):
+        if not self.enabled_user:
+            return
+
         self.process = subprocess.Popen(
             [self.binary, '-L', '--secure', self.rootdir, '--pidfile', self.pidfile],
             stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
@@ -53,5 +64,7 @@ class TFTP(BaseModule):
         pass
 
     def stop(self):
+        if not self.enabled_user:
+            return
         self.process.terminate()
         self.stdout.stop()
